+++ Pulse.BL/Common/Pagination/PaginationDefaults.cs
@@ -2,7 +2,7 @@ namespace Pulse.BL.Common.Pagination;
 
 public static class PaginationDefaults
 {
-    public const int PageNumber = 1;
+    public const long PageNumber = 1;
     public const int PageSize = 10;
     public const int MaxPageSize = 100;
 }
